# backend/app/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HUBMI API — Małopolska Social Innovation Hub",
    "description": "Backend aplikacyjny: matchmaking potrzeb społecznych, innowacje, przepływy pracy, głosowanie i piloty.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/problems/nearby": {
      "get": {
        "tags": [
          "Problemy kanoniczne i Matchmaking"
        ],
        "summary": "Odkrywanie problemów w wybranym promieniu geograficznym",
        "operationId": "get_nearby_problems_api_v1_problems_nearby_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Szerokość geograficzna użytkownika",
              "title": "Lat"
            },
            "description": "Szerokość geograficzna użytkownika"
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Długość geograficzna użytkownika",
              "title": "Lon"
            },
            "description": "Długość geograficzna użytkownika"
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Promień odkrywania w kilometrach",
              "default": 15.0,
              "title": "Radius Km"
            },
            "description": "Promień odkrywania w kilometrach"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CanonicalProblemResponse"
                  },
                  "title": "Response Get Nearby Problems Api V1 Problems Nearby Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problems/{problem_id}/matches": {
      "get": {
        "tags": [
          "Problemy kanoniczne i Matchmaking"
        ],
        "summary": "Dopasowane innowacje społeczne (do 10 wyników)",
        "operationId": "get_problem_matches_api_v1_problems__problem_id__matches_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problems/{problem_id}/coordinates": {
      "get": {
        "tags": [
          "Problemy kanoniczne i Matchmaking"
        ],
        "summary": "Współrzędne 3D dla wizualizacji semantycznej",
        "operationId": "get_problem_coordinates_api_v1_problems__problem_id__coordinates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordinatesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanonicalProblemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "generated_description": {
            "type": "string",
            "title": "Generated Description"
          },
          "reporter_count": {
            "type": "integer",
            "title": "Reporter Count"
          },
          "location_centroid_lat": {
            "type": "number",
            "title": "Location Centroid Lat"
          },
          "location_centroid_lon": {
            "type": "number",
            "title": "Location Centroid Lon"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "recurrence_recommended": {
            "type": "boolean",
            "title": "Recurrence Recommended"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "generated_description",
          "reporter_count",
          "location_centroid_lat",
          "location_centroid_lon",
          "status",
          "recurrence_recommended",
          "created_at"
        ],
        "title": "CanonicalProblemResponse"
      },
...
      "Coordinate3D": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "z": {
            "type": "number",
            "title": "Z"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity_type",
          "title",
          "x",
          "y",
          "z"
        ],
        "title": "Coordinate3D"
      },
      "CoordinatesResponse": {
        "properties": {
          "problem_id": {
            "type": "integer",
            "title": "Problem Id"
          },
          "problem_coords": {
            "$ref": "#/components/schemas/Coordinate3D"
          },
          "solution_coords": {
            "items": {
              "$ref": "#/components/schemas/Coordinate3D"
            },
            "type": "array",
            "title": "Solution Coords"
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "problem_coords",
          "solution_coords"
        ],
        "title": "CoordinatesResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InnovationMatchResponse": {
        "properties": {
          "solution_id": {
            "type": "integer",
            "title": "Solution Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "limitations": {
            "type": "string",
            "title": "Limitations"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "coord_x": {
            "type": "number",
            "title": "Coord X"
          },
          "coord_y": {
            "type": "number",
            "title": "Coord Y"
          },
          "coord_z": {
            "type": "number",
            "title": "Coord Z"
          }
        },
        "type": "object",
        "required": [
          "solution_id",
          "title",
          "description",
          "rank",
          "score",
          "explanation",
          "limitations",
          "coord_x",
          "coord_y",
          "coord_z"
        ],
        "title": "InnovationMatchResponse"
      },
...
      "MatchListResponse": {
        "properties": {
          "problem_id": {
            "type": "integer",
            "title": "Problem Id"
          },
          "total_matches": {
            "type": "integer",
            "title": "Total Matches"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/InnovationMatchResponse"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "total_matches",
          "matches"
        ],
        "title": "MatchListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}